% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_model.R
\name{biasedUrn_rmfnc}
\alias{biasedUrn_rmfnc}
\title{R Wrapper Function for BiasedUrn Sampling}
\usage{
biasedUrn_rmfnc(total, odds, ni)
}
\arguments{
\item{total}{Integer, total number of items to sample}

\item{odds}{Numeric vector of odds for each category}

\item{ni}{Integer vector of population sizes}
}
\value{
Numeric vector of sampled counts
}
\description{
Wraps the BiasedUrn::rMFNCHypergeo function for use in NIMBLE models
}
