% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{create_diagnostic_plots}
\alias{create_diagnostic_plots}
\title{Create Diagnostic Plots}
\usage{
create_diagnostic_plots(chains_list, sim_metadata)
}
\arguments{
\item{chains_list}{List of MCMC chains}

\item{sim_metadata}{Simulation metadata}
}
\value{
List with trace and density plots
}
\description{
Create Diagnostic Plots
}
