% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasymetric.R
\name{dasymetric_mapping}
\alias{dasymetric_mapping}
\title{Dasymetric Mapping}
\usage{
dasymetric_mapping(misaligned_data)
}
\arguments{
\item{misaligned_data}{List with gridx and gridy from simulate_misaligned_data}
}
\value{
sf object with Y grid containing mapped X covariates
}
\description{
Maps X-grid covariates to Y-grid using centroid-based spatial join
}
