% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_model.R
\name{dmfnchypg}
\alias{dmfnchypg}
\title{Density function for multivariate non-central hypergeometric}
\usage{
dmfnchypg(x, total, odds, ni, log = 0)
}
\arguments{
\item{x}{Vector of counts}

\item{total}{Total number of items}

\item{odds}{Vector of odds}

\item{ni}{Vector of category sizes}

\item{log}{Logical, return log probability}
}
\value{
The log-probability (if log=1) or probability (if log=0)
}
\description{
Density function for multivariate non-central hypergeometric
}
