% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasymetric.R
\name{fit_dasymetric_model}
\alias{fit_dasymetric_model}
\title{Fit Dasymetric Model}
\usage{
fit_dasymetric_model(mapped_data, outcome_type)
}
\arguments{
\item{mapped_data}{Output from dasymetric_mapping}

\item{outcome_type}{Distribution type: 'normal', 'poisson', or 'binomial'}
}
\value{
Data frame with parameter estimates and confidence intervals
}
\description{
Fits regression model to dasymetrically mapped data
}
