% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_correlated_spat}
\alias{gen_correlated_spat}
\title{Generate Correlated Spatial Effects}
\usage{
gen_correlated_spat(
  W,
  n_vars,
  rho = 0.6,
  var_spat = 1,
  correlation = 0.5,
  verify = FALSE
)
}
\arguments{
\item{W}{Spatial adjacency matrix}

\item{n_vars}{Number of variables}

\item{rho}{Spatial correlation parameter (default = 0.6)}

\item{var_spat}{Spatial variance (default = 1)}

\item{correlation}{Correlation between variables (default = 0.5)}

\item{verify}{Logical for verification (default = FALSE)}
}
\value{
Matrix of spatial effects
}
\description{
Generate Correlated Spatial Effects
}
