% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.abrm}
\alias{plot.abrm}
\title{Plot method for abrm objects}
\usage{
\method{plot}{abrm}(x, ...)
}
\arguments{
\item{x}{An object of class "abrm"}

\item{...}{Additional arguments (ignored)}
}
\description{
Plot method for abrm objects
}
