% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.abrm_comparison}
\alias{plot.abrm_comparison}
\title{Plot method for abrm_comparison objects}
\usage{
\method{plot}{abrm_comparison}(x, ...)
}
\arguments{
\item{x}{An object of class "abrm_comparison"}

\item{...}{Additional arguments (ignored)}
}
\description{
Plot method for abrm_comparison objects
}
