% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{prepare_adjacency_matrices}
\alias{prepare_adjacency_matrices}
\title{Prepare Adjacency Matrices}
\usage{
prepare_adjacency_matrices(gridy_yorder, gridx_xorder)
}
\arguments{
\item{gridy_yorder}{Reordered Y grid}

\item{gridx_xorder}{Reordered X grid}
}
\value{
List containing W_x and W_y adjacency matrices
}
\description{
Prepare Adjacency Matrices
}
