% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{prepare_nimble_inputs}
\alias{prepare_nimble_inputs}
\title{Prepare NIMBLE Model Inputs}
\usage{
prepare_nimble_inputs(
  bookkeeping,
  adjacency,
  data,
  norm_idx_x = NULL,
  pois_idx_x = NULL,
  binom_idx_x = NULL,
  norm_idx_y = NULL,
  pois_idx_y = NULL,
  binom_idx_y = NULL,
  dist_y = 2
)
}
\arguments{
\item{bookkeeping}{Output from prepare_spatial_bookkeeping}

\item{adjacency}{Output from prepare_adjacency_matrices}

\item{data}{Original simulated data}

\item{norm_idx_x}{Indices of normal-distributed X covariates}

\item{pois_idx_x}{Indices of Poisson-distributed X covariates}

\item{binom_idx_x}{Indices of binomial-distributed X covariates}

\item{norm_idx_y}{Indices of normal-distributed Y covariates}

\item{pois_idx_y}{Indices of Poisson-distributed Y covariates}

\item{binom_idx_y}{Indices of binomial-distributed Y covariates}

\item{dist_y}{Distribution type for outcome (1=normal, 2=poisson, 3=binomial)}
}
\value{
List containing constants, data, and inits for NIMBLE
}
\description{
Prepare NIMBLE Model Inputs
}
