% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{prepare_spatial_bookkeeping}
\alias{prepare_spatial_bookkeeping}
\title{Prepare Spatial Bookkeeping}
\usage{
prepare_spatial_bookkeeping(data)
}
\arguments{
\item{data}{List containing gridy, gridx, and atoms from simulate_misaligned_data}
}
\value{
List of bookkeeping objects for NIMBLE model
}
\description{
Prepare Spatial Bookkeeping
}
