% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.abrm_comparison}
\alias{print.abrm_comparison}
\title{Print method for abrm_comparison objects}
\usage{
\method{print}{abrm_comparison}(x, ...)
}
\arguments{
\item{x}{A abrm_comparison object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for abrm_comparison objects
}
