% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sensitivity_analysis}
\alias{print.sensitivity_analysis}
\title{Print method for sensitivity_analysis objects}
\usage{
\method{print}{sensitivity_analysis}(x, ...)
}
\arguments{
\item{x}{A sensitivity_analysis object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for sensitivity_analysis objects
}
