% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print_convergence_summary}
\alias{print_convergence_summary}
\title{Print Convergence Summary}
\usage{
print_convergence_summary(convergence_results)
}
\arguments{
\item{convergence_results}{Results from check_mcmc_diagnostics}
}
\description{
Print Convergence Summary
}
