% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_model.R
\name{Rmfnchypg}
\alias{Rmfnchypg}
\title{Nimble R Call Wrapper for BiasedUrn}
\usage{
Rmfnchypg(total, odds, ni)
}
\arguments{
\item{total}{Total number of items}

\item{odds}{Vector of odds}

\item{ni}{Vector of category sizes}
}
\value{
Vector of sampled counts
}
\description{
Internal wrapper to call R function from compiled Nimble code.
}
\keyword{internal}
