% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.abrm}
\alias{summary.abrm}
\title{Summary method for abrm objects}
\usage{
\method{summary}{abrm}(object, ...)
}
\arguments{
\item{object}{An abrm object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for abrm objects
}
