% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{waic}
\alias{waic}
\title{Generic Function for WAIC}
\usage{
waic(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{"waic_abrm"} object for \code{abrm} models. See
  \code{\link{waic.abrm}} for details.
}
\description{
Extracts the Widely Applicable Information Criterion (WAIC) from a
fitted model object.
}
