% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.abrm}
\alias{plot.abrm}
\title{Plot method for abrm objects}
\usage{
\method{plot}{abrm}(x, ...)
}
\arguments{
\item{x}{An object of class "abrm"}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object \code{x}. The function is called for its side effect of displaying MCMC diagnostic plots (trace plots and density plots) if they are available in the abrm object.
}
\description{
Plot method for abrm objects
}
