% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.abrm_comparison}
\alias{plot.abrm_comparison}
\title{Plot method for abrm_comparison objects}
\usage{
\method{plot}{abrm_comparison}(x, ...)
}
\arguments{
\item{x}{An object of class "abrm_comparison"}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object \code{x}. The function is called for its side effect of displaying comparison plots between methods if true parameters are available in the comparison object.
}
\description{
Plot method for abrm_comparison objects
}
