% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.abrm}
\alias{print.abrm}
\title{Print method for abrm objects}
\usage{
\method{print}{abrm}(x, ...)
}
\arguments{
\item{x}{An abrm object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for abrm objects
}
