% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.misaligned_data}
\alias{print.misaligned_data}
\title{Print method for misaligned_data objects}
\usage{
\method{print}{misaligned_data}(x, ...)
}
\arguments{
\item{x}{A misaligned_data object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for misaligned_data objects
}
