% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.vcov.abrm}
\alias{print.vcov.abrm}
\title{Print method for vcov.abrm objects}
\usage{
\method{print}{vcov.abrm}(x, ...)
}
\arguments{
\item{x}{A vcov.abrm object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for vcov.abrm objects
}
