% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_model.R
\name{register_nimble_distributions}
\alias{register_nimble_distributions}
\title{Register Custom NIMBLE Distributions
Registers the custom distributions for use in NIMBLE models.}
\usage{
register_nimble_distributions()
}
\value{
Invisible TRUE
}
\description{
Register Custom NIMBLE Distributions
Registers the custom distributions for use in NIMBLE models.
}
\keyword{internal}
