% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_model.R
\name{rmfnchypg}
\alias{rmfnchypg}
\title{Random generation for multivariate non-central hypergeometric}
\usage{
rmfnchypg(n, total, odds, ni)
}
\arguments{
\item{n}{number of observations (only n=1 is used)}

\item{total}{Total number of items}

\item{odds}{Vector of odds}

\item{ni}{Vector of category sizes}
}
\value{
Vector of sampled counts
}
\description{
Random generation for multivariate non-central hypergeometric
}
\keyword{internal}
